chore(basic-auth): moved email and password auth into basic directory under auth
