+++ src/auth/auth.service.ts
@@ -6,9 +6,9 @@ import { Response, Request } from 'express';
 import * as bcrypt from 'bcrypt';
 import { User } from '@prisma/client';
 import { DatabaseService } from '@database/database.service';
-import { LoginDto } from './dto/login.dto';
-import { CreateUserDto } from './dto/create-user.dto';
-import { JwtPayload } from './types';
+import { LoginDto } from 'src/modules/auth/dto/login.dto';
+import { CreateUserDto } from 'src/modules/auth/dto/create-user.dto';
+import { JwtPayload } from '../modules/auth/types';
 import { API_CONSTANTS, AUTH_LITERALS, CONTENT_TYPES } from '@config/constants';
 import { ConfigService } from '@nestjs/config';
